#8 교수용 로그인 회원가입 기능 구현 완료 (#18)
fix #8
