+++ pubspec.yaml
@@ -64,6 +64,9 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
+  assets:
+    - asset/image/JST.jpg
+
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
